Don't rely on there being a style element already present in the document; instead, create one if it doesn't exist.
The reason for doing it this way is that we are no longer use UIWebView's loadHTMLString
method to create the preview page with style element, which was done in StyleManager's
initialisation routine. loadHTMLString is asynchronous, meaning we could not immediately
display a preview. Now, we can set the style to "Normal" during initialisation, and the
preview will be updated correctly, because the javascript code will just create the style
element.