Removed Range.convertToOffsetFree(). While we can safely avoid using start offsets, we need to use end offsets because otherwise there is no way to represent the case where the selection ends before the first child of a node.
For example, consider the case where the start and end of the selection are at the
location indicated by * below:

<div>
  *
  <div></div>
  <div></div>
  <div></div>
</div>

With the "offset free" approach, where the end node is an element, the range is assumed
to end after the last child of the element. Previously, we simply moved the end node to
the previous node, but this could cause it to move before the start node, become the
parent node, or if it's right at the end of the beginning of the document, become null.

We will simply have to deal with offets that are relative to element nodes.
getSelectedNodes() may need to be updated to account for this fact.