Changed isFirstInParagraph() to ignore whitespace text nodes preceding the node in question.
Restructured nodeHasContent() and acceptNode() to hilight the similarities in the first
part of both functions.

When backspace is pressed at the start of a paragraph, only add the contents of the
current paragraph if it has content. Also, remove a trailing BR at the end of a paragraph
when backspacing into it.