Support for maintaining a position while a node is moved. Previously, we only knew about separate removal and insertion events, with no way to correlate them, and thus if a note was moved (which involves removal followed by insertion), the position could not be automatically adjusted.
Now, when a node is moved using the moveNode method, any postions which are located
directly preceding that node will be updated so that they still directly precede the node
in its new location. This is achieved by temporarily changing the position to (node,0)
before the removal occurs, and then afterwards changing it to (node.parentNode,offset)
once it is inserted again. The removal and insertion handlers in the Position class
identify nodes that are in the process of being moved via the "moving" attribute, which
moveNode temporarily sets while the move is taking place.