Added movePreceding and moveFollowing, which are similar to the *SiblingsToOtherNode methods except they use (node/offset) pairs rather than just nodes, thus supporting the case where there is no node at the current position. (node/offset) pairs are more appropriate here, because we want to move all nodes before or after a given *point*.
Modified selectionWrapElement, selectionUnwrapElement, and splitAroundSelection to work
with ranges that use (node,offset) positions.

Changed several places where nodes are moved from one location in the tree to another to
use moveNode, which causes any tracked positions to be updated. The three methods
mentioned above also now use trackWhileExecuting to make sure the selection start and end
are adjusted to take into account the modifications they make to the DOM tree. Merging of
text nodes is not yet handled correctly with respect to this (which causes the
wrap-multiplep3 test to fail).

Modified mergeRange to use getAllNodes rather than going through the selection using
nextNode, a technique which doesn't work with (element,offset) positions.