Added exhaustive tests for isForwards() and getOutermostNodes(), which run a very simple (but inefficient) version of the algorithm on every combination of start and end positions in an sample tree against the regular (complex but more efficient) methods of the Range class.
As a result of this testing, a number of bugs were found in both methods that only
occurred on obscure corner cases that had not been previously tested. The methods have
been updated so that work correctly and pass all the tests.

On particular, one significant change to getOutermostNodes() is that it now returns an
empty array if there is no content between the start and end positions, including the
case where the start and end position are part-way through a text node. An implication of
this is that when the selection is just a cursor, there is nothing actually "selected",
and getOutermostNodes() will return nothing. applyFormattingChanges() has been partially
modified to take this into account.