Renamed markOutlineDirty() to scheduleUpdateSectionStructure()
updatSectionStructure() now checks the dirty flag and if it's false, doesn't do anything.
Otherwise, it clears it, and proceeds to do its work. This way, if updateSectionStructure
happens to be called explicitly when the is already a postponed action in the queue,
it won't repeat its work a second time.

Modified moveSection() to call updateSectionStructure() first, to make sure all the
pointers between sections (parent, children, prev, next) are valid, because the algorithm
for moving sections relies on these.