Exposed mergeWithNextSibling() and nodesMergable() as public functions
Changed mergeWithNextSibling() so it makes a recursive call to itself at the end (if
appropriate), instead of calling mergeWithNeighbours(). There are situations in which we
only want to match the old last child of the current node with the old first child of the
next node, e.g. when deleteting the selection contents.