Better solution to the issue I tried to address in revision 1100, but ended up breaking lots of test cases.
Previously, a cursor position that was not adjacent to a character or a "content" element
like an image or table was considered invalid. However, when the user is typing and they
press space, we need to consider the position after the space as valid (assuming it is at
the end of the line, otherwise we will move to the start of the next character).