Got rid of the DOMCharacterDataModified listener in Position.js, and force all changes to the nodeValue of text nodes to be made through the newly-added DOM functions insertCharacters(), deleteCharacters(), and setNodeValue(), which adjust the positions in an appropriate manner for the type of change.
The old listener code only knew that there was a change, not whether it was an insertion,
deletion, or replacement. As a result, it had to try and be clever about figuring out
which of these types of changes had occurred, and wasn't always able to get it right.
For example, if you have the string "aaabbb" and you notice that it changes to "aaabb",
you have no way of knowing which of the bs was deleted. One of the implications of this
is that when the cursor was after a sequence of two or more characters that were the same
and the user pressed backspace, the cursor would move to the wrong location.

Now that we are using the insert/delete/set methods mentioned above, we can adjust
positions accurately.