Show the cursor menu in the correct situations, i.e. after the user has tapped in the same cursor location twice (which was already working), and after they have dragged the loupe around.
Modified the JS positionCursor() method to return a true or false value indicating whether
or not the cursor is in the same position as in the previous call. When the user taps the
editor view, the result of this method is used to decide whether or not to display the
cursor menu. Previously, positionCursor() called showCursorMenu(), which we *don't*
want to do when the user is dragging the loupe around - in this latter case, we only want
to show the cursor menu once the drag operation has finished (i.e. they have lifted their
finger from the screen).