Replaced calls to DOM.insertBefore() and DOM.appendChild() within setList() to use DOM.moveNode() instead, so that positions are correctly tracked. DOM.moveNode() now checks to see if the node is already in the tree (i.e. if its parent is non-null), and if so, calls DOM.insertBefore()
The reason for this change is that sometimes in setList() we are moving existing nodes,
and sometimes we are adding new nodes, and it's not obvious from the code which of these
two cases is taking place in a given situation. This is actually a more general problem
with the DOM class - we should be using one method universally, rather than having
two different methods for the new/existing node case.