First steps towards moving stylesheet data structures into JS (which will be important for future non-destructive editing of word documents)
Moved getStyles() into Styles.js, now called reportStylesToEditor(). Added
Styles_count() and Styles_styleAtIndex().

Ran into a bit of a roadblock in JSInterface.m. I want to use the Formatting class, but
there's a property of JSInterface which has the same name, causing a conflict. I need to
change the properties of JSInterface to start with a lowercase letter before I can proceed
(there are others like Outline as well).