Translated Formatting.cssTextForSelector into javascript, as condenseProperties() and propertyListText() in Styles.js
Added a mechanism whereby app-default style properties such as table borders can be added,
assuming the author has not previously set that property to something else in the
style sheet.

When inserting tables, don't explicitly specify the border property. Instead, rely on
the default style rule to set the CSS border properties for TD elements.

When inserting tables, set the column widths to be equal, based on the appropriate
percentage of the table width (based on the number of columns).

Added a couple of tests for basic table insertion.