When re-generating the text of a stylesheet in applyCSSTextChanges(), group rules that have identical property values together by separating the selectors with commas.
Add the default styles *before* examining the document's stylesheet to discover explicitly
defined styles. This is because we set some styles as latent when creating the default,
and we want to change them to be non-latent if we find that they're already present in
the document.

In discoverStyles(), split the selector text into individual selectors (which are
separated by commas), to be compatible with the form of rules we generate above.
This isn't a strictly perfect solution, because we don't correctly handle selectors
containing commas inside quotes, e.g. h1[foo="bar,baz"]; doing so would require properly
parsing the selector according to the CSS grammar.