Fixed a case in which pressing backspace - i.e. calling deleteCharacter() - would timetimes result in a null pointer dereference. This was being caused by the selection being incorrectly set because the offset was placed at the end of a selection display DIV.
I added the hideSelection() function, which is called at the start of
deleteSelectionContents(), to make sure these DIVs aren't present when we're adjusting
the selection in this function. However, this needs to be done more generally - there
should be a hideSelection() call before any operation that modifies the selection, and
a showSelection() after.