Deprecated setSelectionRange() and setEmptySelectionAt(). The preferred approach now is to call Selection_hideWhileExecuting(), passing a function to be executed, and make any changes in that, calling Selection_set() or Selection_clear() to modify the selection, which will be displayed again when the operation is complete.
The reason for this change is that I'm going to attempt to change the way the selection
is displayed by using a call to applyFormattingChanges() to set the background of the
selected range of text to light blue, rather than using DIVs. This would result in the
light blue selection background being visible *on top of* any paragraph backgrounds,
in contrast to the current issue where a paragraph that has a background is displayed
over the selection DIVs, since they have a z-index of -1.