Changed the way in which we implement hooper selection. Now, instead of recognising the gesture when the user drags their finger a certain minimum distance over the keyboard, we have special "move" and "select" buttons in the accessory keyboard. You have to hold down one of these to activate the mechanism.
In Daniel's original video, it was possible for the user to do this without having to
explicitly press a button, but I found that two-finger gestures conflicted with iOS's
built-in split keyboard mechanism, and there were also some potential issues with single-
finger dragging due to the ability to press and hold on a key to bring up alternatives.
Apparently some non-english keyboards also rely on gestures.

Requiring the user to hold down a button while moving the cursor eliminates these hassles,
as we can safely dim the keyboard and take over the whole view for our own purposes. It
also avoids potential issues with acceptance on the app store in which Apple could claim
that it interferes with built-in iOS functionality.