Added tests for changing formatting properties when the selection is empty (i.e. just a cursor). We need to make sure that when this happens, any subsequently inserted characters will have this formatting applied.
Added an additional parameter to isValidCursorPosition(), 'insertion', that indicates that
we want to know whether it is a valid place to insert a character, rather than just a
valid place to use when using the left and right arrow keys. This is slightly less strict,
in that it allows the cursor to be placed directly after an inline element like <i>. When
applying formatting changes, we set this parameter to true, so that for example if the
user switches italics off, then the next character they type will not appear in italics.