Part 1 of preparations for Selection_get() restrictions
These tests were calling Selection_get() and then passing the result to getFormatting(),
which was unnecessary, as getFormatting() does not take any parameters (i.e. it ignores
the range that was being passed in), and instead determines the range to examine by
calling Selection_get() itself.

All changes replaced the following:

    var range = Selection_get();
    var properties = Formatting_getFormatting(range);

with:

    var properties = Formatting_getFormatting();