Part 7 of preparations for Selection_get() restrictions
Got rid of the PostponedActions_performImmediately flag, which was used by some testcases
to avoid havoing to explicitly call PostponedActions_perform() at the necessary points
during execution. With the upcoming introduction of the requirement that the selection
can only be accessed while it is hidden, this would result in nested calls to
Selection_hideWhileExecuting(), which are not allowed.