Imposed the restriction that Selection_get() can only be called while the selection is hidden. The reason for this is as follows:
In order to support proper placement of the cursor in the case where center or right
alignment is used and the cursor is after a space, we are going to have to place a
non-breaking space (&nbsp; or character 160) between the last non-whitespace character
and the cursor position. Without doing this, the current technique of temporarily
inserting an 'X' before the cursor position when calculating its location causes the text
to be shifted left, but only while the calculation takes place - and when the text is
shifted back (after removing the 'X'), the cursor position ends up being further to the
left than it should be.

Adding the &nbsp;, and thus shifting the text to the left, is also necessary to give the
expected visual feedback when the user types a space at the end of a line. Without this,
the horizontal position of the line would change when typing non-whispace characters, but
not when typing a space.

In imposing this restriction, we are effectively saying that the DOM can be in one of two
modes - "presentational" mode, or "actual" mode. In presentational mode, certain
modifications are made so that the cursor or selection looks right on screen, though these
changes aren't considered part of the document proper, and won't be persisted when saved.
This allows us to make arbitrary modifications to the DOM while presenting the cursor or
selection (including, in the future, adding extra spans to set the background colour),
while preventing all code that actually implements editing operations from actually seeing
these temporary changes.

The &nbsp; hack described above is yet to be introduced; this is just preparation for it.