Instead of using absolutely-positions DIV elements to reperesent the blue selection highlight, use SPAN elements wrapped around every text node that is part of the selection.
This is necessary because with the former technique, if an element had a background colour
set, the selection highlight would not show as it would appear behind the element's
background. With the latter technique, the SPAN background is always drawn on top of the
parent's background; the SPANs are always the furthest down in the tree (directly above
the text nodes), so their background takes precedence over any other (including inline
backgrounds set on other SPANs).

In the hide() method, we no longer get rid of the seleciton highlights, but instead just
leave them there and update them the next time show() is called. This represents a step
away from the hide/modify/show model we've been using lately, and will hopefully enable
us to abandon this entirely in the future. Once I've sorted out the space at end of line
issue with center/right-aligned paragraphs, we should be able to always work with the
selection range as-is, without having to temporarily hide it while doing modifications.

Leaving the SPANs in place and simply updating them when the selection changes also
improves performance (especially when large amounts of text is selected), because it means
we no longer have to destroy and re-create all the selection highlight elements every time
the selection changes. This is particularly important when using drag-to-select, as it
makes each upate much faster.