Properly position the cursor with respect to the caption when moving around a table.
This is needed because HTML requires that the <caption> element be the first child of a
<table>, though we use the css propery caption-side: bottom to display the caption below
the table, rather than above (for consistency with LaTeX). Considering the table caption
in document order, as we did up until now, meant that if you used the left and right keys
to move in and out of a table, it would go into the caption first (when moving right)
instead of the cells. Now the caption is treated as if it comes after the cells in
document order.