Bugfix: Crash on pressing enter when cursor is directly after an autocorrect span
Cursor_enterPressed() uses Formatting_splitTextAfter() to create a copy of the containing
elements and insert them in the new paragraph. The autocorrect span was being inserted
twice (once on the span split, a second time on the paragraph split); this in itself
wasn't a problem. However, a new Correction was created each time (for the same span
element), and the program subsequently tried to remove both of them, effectively removing
the same span twice.

Although there was a done variable in the Correction constructor to prevent against a
given correction being removed multiple times, it didn't take into account the possibility
of multiple correction objects associated with the same span each being removed. Now,
instead of this, we just check in removeCorrectionSpan() whether the span has a parent
node or not.