Added Text.js, with code to analyse a paragraph and produce a series of runs. Each run corresponds to a text node, and records the start and end offset of that node within the total string of the paragraph.
This is going to be useful for a whole bunch of things. I've initially added it for
supporting some of the keyboard shortcuts with the extended keyboard (such as moving to
the beginning or and of a word), but it will also come in handy for find & replace, spell
checking, and translating between HTML and DOCX's concept of runs.

Also added skeleton functions for all the keyboard shortcut-based cursor movement
operations, as well as a mostly-complete implementation of moveToStartOfWord()