Use Position_okForMovement instead of Position_okForInsertion as the criteria for determining the next/previous position to check in posAtEndOfLine() and posAtStartOfLine()
It seems that WebKit in Safari (5.1.7) allows us to get the client rect for positions
in text nodes that have whitespace on either side, but WebKit on iOS (5.1.1) doesn't.
So by using Position_okForMovement we avoid checking this positions. Prior to this
chane, if you pressed command-left or command-right on the iPad, in some cases the cursor
would only go part way along the line, because there happened to be multiple spaces in
the actual text node.