When enter is pressed, don't split more than the current paragraph.
This solves a major performance problem with HTML files exported from MS word, which
basically look like this:

<html>
  <body>
    <div>
      <p>...</p>
      <p>...</p>
      <p>...</p>
      ...
    </div>
  </body>
</html>

Every time the user pressed enter, it would create a new div element, and move all of the
following paragraphs into this div element. On a large document, where enter was pressed
near the start of the file, this could take a long time (in the order of seconds).

Now, in this situation we just split the current p element, and leave it and everything
else inside the containing div.