In Position_displayRectAtPos(), avoid ever adding a temporary node or character to the DOM to determine a position, because this causes a relayout and is really slow on large documents. Instead, if we can't find a rect for the exact position requested, search backward or forward (only within the same rect).
If we still can't find a position, return null. The callers of this function now take
no action if a rect can't be found, which I think is better than the other two options of
using the wrong position, or throwing an exception. The only current exception is
AutoCorrect_getCorrectionCoords(), which still needs to be fixed (this requires handling
on the Objective C side).