Perform the removal of special UX Write-specific elements and attributes as part of the HTML tidying process, in C, rather than the Main_getHTML() function in javascript.
This is both more efficient, and - more importantly - allows us to do this removal process
in a separate thread (which is important for minimising user interuption during autosave).

Some rough benchmarks on my thesis (note the Total includes other short stages in addition
to JS getHTML/HTML Tidy) (average over 3 runs)

BEFORE
    JS getHTML: 1986
    HTML Tidy: 2712
    Total: 4796

AFTER
    JS getHTML: 207
    HTML Tidy: 3248
    Total: 3549

Saving 26%

And note in particular the saving in the portion of the code that requires us to block
the main UI thread - 89% (207ms, down from 1986ms).

This means that if we also implement a delay to make sure the user has been idle for
a little while, the user likely will never notice the autosave process.