Fixed a crash in Input_replaceRange() in the case where the position was an element position rather than a text position. This could be observed by starting a new paragraph, and entering one of the system's built-in text shortcuts (such as "omw" for "On my way!").
The problem was due to the fact that I was using incorrect variables for accessing the
node used in insertBefore.

This is actually the first time I've figured out what this method is for; I wasn't really
sure what situations it was used in before, which explains why I hadn't caught this bug
previously. A bug report by ialyabushev@flexis.ru (19 August) gave detailed instructions
on how to reproduce the problem.