In applyFormattingChanges(), apply the changes to the *innermost* paragraphs, rather than the outermost paragraphs.
This is important for situations like HTML exported from word, where you have an outer DIV
element containing all of the paragraphs, and you want to modify the paragraph formatting
individually, without touching the DIV (which in this example would alter the entire
document).

Related to this, also changed ensureValidHierarchy() to allow heading elements inside
DIVs.