Removed the Input_beginningOfDocument() and Input_endOfDocument() javascript functions. Instead, from EditorInput, we now return the single abstract position we maintain for these two.
Also changed the selectedTextRange method to return the single selection range that we
maintain, similar to the document range used for the above-mentioned beginning and end
positions. As long as the text input system has not changed the selection (via
setSelectedTextRange:), we return the same selection object which has abstract positions
corresponding to the start and end of the selection. This means we also avoid javascript
calls when the selectedTextRange method is called.

When setSelectedTextRange: is called, we convert the abstract positions from the current
selection range into concrete ones, as the text input system sometimes keeps then around
and reuses them in later calls (e.g. when asking about the position n lines up or down
from what was previously selected). In this case we can't simply rely on using the
current selection start/end in place of these positions, because it assumes those
positions represent the same location in the document as they did before.

Also changed positionFromPosition:offset: to avoid a javascript call by creating a new
Position object with the appropriate adjustment.