Introduction of type ids part 2: In DOM_assignNodeIds(), set the _type field of the node based on the value from the ElementTypes hash table.
Increases time for initial DOM_assignNodeIds() call from to 200ms on my thesis
(tested on iPad 3, iOS 6 beta 4). Note there are 32,759 nodes in this document.

This changes brings us back to about where we were before the optimisations that changed
to the use of numeric node ids, and avoiding the check for existing ids on each call.
However the type ids give us huge benefits in other areas, as we can compare by numeric
type rather than strings (which until now we've always been converting to uppercase, even
if they don't need to be).