Bugfix: Incorrect heading numbering when a TOC is placed after one or more headings
This was caused by the fact that when you insert a new heading, it searches backwards in
the document to find the previous heading, and returns the outline item object associated
with that to use as the predecessor in the list of all sections. If the preceding heading
happened to be the heading of a TOC, there was no associated item, so prevItem in
Category_add() would be null. This caused the new heading to be added to the start of the
list, and get the number 1, while the other headings earlier in the document would get
the numbers after this.