Word: When inserting a table in the editor, create and use the TableGrid style, so that we have borders.
This necessitated the modification of the javascript insertTable() method so that it
always places the tr elements inside a tbody, as the selector format we are using to
match table cells requires this. WebKit always adds this when parsing HTML, but this
wasn't previously done when adding a table directly using insertTable().