When the stylesheet has changed, send both the serialised CSS text and also a JSON representation of the stylesheet to the javascript Styles module. This way, we have an easy-to-access representation of the entire stylesheet in javascript code, and don't have to rely on the DOM CSS API or use getComputedStyle() (which is buggy, at least in the case of counters).
This facility isn't actually used for anything yet, but it will be when we implement
automatic numbering using CSS counters, since we'll need to examine the style rules to
determine if automatic numbering is actually in effect.

To simplify updating the stylesheet contents from the Objective C side, we now have two
new methods: EditorWebView.updateCSS:, which takes a StyleSheet object as a parameter, and
EditorViewController.updateCSS, which updates based on the editor's own stylesheet
object (which the EditorWebView object does not have a reference to).