Replaced the javascript-based calculation and setting of heading, figure, and table numbers with CSS counters. This is a better approach because it allows customisation of whether numbering is enabled, and in what format, as part of the stylesheet - and in a manner that follows the mechanisms built in to the standard, rather than the "roll your own" solution we were using before.
For backwards compatibility, the first time we open a document that doesn't use this
scheme, we scan through it to detect the presence of heading, figure, and table numbers
in the same manner as before. These are all removed, and the appropriate elements changed
to use the appropriate style.

For headings, if there are one or more headings found that begin with a number, then we
set the default numbering style for h1 through h6 to numbered (in x.y.z format). If
this is the case, then any headings which *don't* begin with a number have their class
set to "Unnumbered" (while still remaining as hX elements). Alternatively, if there were
no headings found that start with a number, then the default heading styles are left
alone, and "Heading Numbering" in the settings menu will appear as off.

This is a one-time conversion process that applies only to documents created with UX Write
version 1.0.x, or HTML documents created by hand or by other applications. Once we have
done this conversion, we set the counter-reset property of the body element to reset the
counters h1, h2, h3, h4, h5, h6, figure, and table. When opening a document, the program
checks the counter-reset property of the body element, and if it contains *anything*, it
assumes that CSS counters are in use, and does not carry out the conversion process.

This commit breaks the following things:
- Ability to turn numbering on or off for figures and tables
- Almost all of the javascript outline test, plus some others that depend on the old
outline functionality