Remove .Unnumbered styles if heading numbering off
When heading numbering is on, a set of styles h[1-6].Unnumbered are
defined, corresponding to \section* and friends in LaTeX. If the user
turns of heading numbering globally, we don't need these, so remove them
from the stylesheet, and also update all the heading elements in the
document to clear their class attributes.

This is particularly important in the case of Word documents, where you
see the "Heading X (Unnumbered)" styles in the style list if they're
there, which is a bit unsightly.