Tables: Correctly handle pasting cells in a table
When selection is inside a table, and a paste is performed whose content
is itself a table (or row or cells), replace the cells starting from the
top-left most cell of the selection with those from the clipboard.

The case where merged cells exist in the pasted-into table is handled by
splitting them all before the paste.

The case where merged cells exist in the pasted-from table is handled by
simply pasting them as-is, making the appropriate adjustments to the
size of the pasted-into table as necessary.

If there are more rows or columns in the clipboard, the pasted-into
table is extended with additional rows and columns as necessary,
including empty cells that are not within the region that will be pasted
to.