Fix exception in select all when table selected
If a select all action is performed, we try to set the selection range
range to (document.body,0,document.body,document.body.childNodes.length)
which was causing problems if there were a range of table cells
selected.

An exception would occur when Selection_update() removed the table
selection divs (the last children of the body), and then set the offset
of the selection range end to what has then become an invalid offset.

Now, when we remove the table selection divs, we do so within the
context of Range_trackWhileExecuting(...). This ensures that the end
of the selection is moved backwards when the table selection divs are
removed, resulting in a valid selection range after the update is
performed.

Reported by Lisa Harvey on 12/3/2013