Tables: Avoid duplicate ids on cell paste
When copying & pasting cells within or between tables, we need to make
sure that we don't end up with duplicate id attributes, because this
confuses the BDT algorithm, causing it to omit content from the saved
file. This was already done in the Clipboard_pasteNodes() function using
removeDuplicateIds(), but now we call this at the end of pasteTable() as
well.