Multi-stage text input (Chinese/Japanese)
Modified EditorInput (and corresponding javascript functions) to support
the setMarkedText and unmarkText methods, which are needed for languages
such as Chinese and Japanese that use a multi-stage text input
mechanism.

Rather than maintaining separate ranges for marked text and selected
text, we simply just use the selection for one. Even though the
setMarkedText method takes a parameter that specifies which of the
characters within the range are supposed to be part of the selection,
this range always seems to be empty, and placed at the end (or
NSNotFound). So we just ignore that parameter.

The Selection module in javascript has also been modified to keep track
of a flag that indicates whether the current selection is marked text or
not. If it is, the blue selection highlight is still shown, but not the
selection handles.