Fix cursor placement beyond start/end of body
In Position_atPoint(), we check to see if the cursor is above the top of
the body, or below the bottom of the body. If so, we search forwards or
backwards (respectively) for the closest suitable position in which to
place the cursor. This is done so that if you have text on the first or
last line, you can still move the cursor about by tapping in the area
above or below the appropriate character on a line.

However, this caused undesirable behaviour in the case where the
document started or ended with an empty paragraph - tapping outside the
body area would *not* allow the cursor to be placed inside that
paragraph, because it did not have any content. Now, we check for empty
paragraphs, and if we find one, we place the cursor there instead of
continuing to the boundary of the nearest character.