Remove Outline_resetSectionNumbering()
This function was called after turning heading numbering on or off. It
went through the document and cleared the class attribute on headings.
The idea behind this was that previously we were only using heading
class attributes to distinguish between numbered (no class) and
unnumbered ("Unnumbered" class). Now, with all styles having a class,
this no longer makes sense.

Instead, we just call Outline_scheduleUpdateStructure() after turning
heading numbering on or off, so it can update the cross-references
appropriately.