Fix extra character bug at start of japanese word
Cursor_insertCharacter() checks to see if the position is a valid one
for insertion, moving forwards or backwards as appropriate if it isn't.
In a bug reported by Nishi Hiro (27 August), if you pressed space, then
started typing in the letters for a japanese word, the first (english)
letter would remain at the start, as well as being considered part of
the marked text.

This commit makes two fixes, both within Cursor_insertCharacter()

- If the position is moved to a valid insertion point, and the range
  between the original position and the new one has no content, that
  range is deleted. This gets rid of the extra space in front of the
  cursor in the case where the position is moved back one character.

- Attempts to insert a space at the beginning of a paragraph are ignored