decreaseIndent: make top-level list items <p>
Change the behavour of decreaseIndent so that when this function is
called with a selection that includes top-level list items (i.e.
non-nested items), then those become regular paragraphs. This behaviour
seems a bit more intuitive to me, and makes it a bit easier for working
with lists, as it's now possible to change from a (non-nested) list item
to a paragraph just by using the decrease indent button/keyboard
shortcut.