Bugfix: Merging paragraphs after backspace
Fix a bug in which if you were at the end of a paragraph, typed a
misspelt word and pressed space, it was autocorrected, and then you
pressed backspace, the following paragraph would be merged into the
current one.

This was particularly annoying in the case where the following paragraph
had a different style, such as a heading, as you couldn't just fix the
problem by pressing enter (you had to change the style back).