Fix getFormatting09 test in Chrome
When you set the border property in a CSS style definition or HTML style
attribute, Chrome actually sets several border-image-* properties as
part of that process, and these show up when you enumerate the list of
properties. This behaviour does not occur in Safari. The test was
already ignoring the "border-image" property that was used in previous
versions of chrome, but now it instead checks for properties that begin
with the string "border-image-" and ignores all of them.