Fix cursor visibility after inserting a note
This commit fixes a bug where after inserting a footnote or endnote, and
then moving the cursor to the position directly after it, the cursor
would show up as being inside the brackets, rather than outside.

This was happening because in some cases (such as when the cursor is
part-way through a text node at the time of the insertion), we would end
up with an empty text node appearing directly after the note. The empty
text node came from Formatting_splitAroundSelection.

When the arrow keys were subsequently used to move the cursor past the
last character in the note, it would go into this empty text node.
Position_rectAtPos would not be able to find the proper coordinates for
the empty text node, and thus it would look as though the cursor was
still inside the note, even though it was not.

Now, before doing the actual insertion of a note, we remove the empty
text node, if it's present.

The insertFootnote and insertEndnote tests have now been updated to call
showEmptyTextNodes, but the expected result remains unchanged - since we
should not see any empty text nodes after insertion.