Fix disappearing footnotes/endnotes on formatting
Fix a bug that caused footnotes and endnotes to be converted back to
normal text when formatting was applied to any part of the paragraph,
either inside or outside the note.

This was caused by removePropertiesSingle looking at the span element,
seeing that it had no style set, and assuming that meant it could remove
it safely.

Actually, there's technically another bug here, in that if a span has a
class name, it shouldn't be removed at all. When we add support for
character styles, we should fix this as well.