displayRectAtPos: Fix result on note boundary
If we're immediately before or after a footnote or endnote, calculate
the rect by temporarily inserting a space character, and getting the
rect at the start of that. This avoids us instead getting a rect inside
the note, which is what would otherwise happen if there was no adjacent
text node outside the note.