Remove trace() wrapper from javascript code
iOS 5 (and possibly 6, not sure) lacked a stack property in exception
objects, and so to capture this information we had a trace() function
which would wrap each function and add the necessary information so we
could get a proper call stack with function names.

iOS 7 provides a stack property for exceptions, so we can just consult
this instead. Thus, the trace() function is no longer necessary. It was
quite inconvenient using this as you had to remember to put it around
every function (if you wanted it included in stack traces), and you
couldn't define functions using the normal syntax.

While we still use variable assignment in many places (to make functions
within a closeure accessible globally), we no longer do this for private
functions, and even public functions are assigned directly to a global
variable, instead of being passed to trace().

Another nice thing we get in iOS 7 is the line & column number at which
the exception occurred. Now that we're using the stack trace provided
natively by the JS interpreter, we can include this in error reports.